+++ ccguard/__init__.py
@@ -5,4 +5,7 @@ from .ccguard import (  # noqa
     configuration,
     GitAdapter,
     adapter_factory,
+    parse_common_args,
+    print_cc_report,
+    print_delta_report,
 )
